pyproject.toml
[project]
name = "us"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.4",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=25.3.0",
    "psycopg2-binary>=2.9.12",
    "whitenoise>=6.12.0",
]
